remove shared options for sse
