+++ server/sse.go
@@ -36,6 +36,13 @@ type sseSession struct {
 // content. This can be used to inject context values from headers, for example.
 type SSEContextFunc func(ctx context.Context, r *http.Request) context.Context
 
+// DynamicBasePathFunc allows the user to provide a function to generate the
+// base path for a given request and sessionID. This is useful for cases where
+// the base path is not known at the time of SSE server creation, such as when
+// using a reverse proxy or when the base path is dynamically generated. The
+// function should return the base path (e.g., "/mcp/tenant123").
+type DynamicBasePathFunc func(r *http.Request, sessionID string) string
+
 func (s *sseSession) SessionID() string {
 	return s.sessionID
 }
@@ -93,7 +100,7 @@ type SSEServer struct {
 	sseEndpoint                  string
 	sessions                     sync.Map
 	srv                          *http.Server
-	contextFunc                  HTTPContextFunc
+	contextFunc                  SSEContextFunc
 	dynamicBasePathFunc          DynamicBasePathFunc
 
 	keepAlive         bool
@@ -102,41 +109,37 @@ type SSEServer struct {
 	mu sync.RWMutex
 }
 
-// Ensure SSEServer implements httpTransportConfigurable
-var _ httpTransportConfigurable = (*SSEServer)(nil)
+// SSEOption defines a function type for configuring SSEServer
+type SSEOption func(*SSEServer)
 
-func (s *SSEServer) setBasePath(basePath string) {
-	s.basePath = normalizeURLPath(basePath)
-}
-
-func (s *SSEServer) setDynamicBasePath(fn DynamicBasePathFunc) {
-	if fn != nil {
-		s.dynamicBasePathFunc = func(r *http.Request, sid string) string {
-			bp := fn(r, sid)
-			return normalizeURLPath(bp)
+// WithBaseURL sets the base URL for the SSE server
+func WithBaseURL(baseURL string) SSEOption {
+	return func(s *SSEServer) {
+		if baseURL != "" {
+			u, err := url.Parse(baseURL)
+			if err != nil {
+				return
+			}
+			if u.Scheme != "http" && u.Scheme != "https" {
+				return
+			}
+			// Check if the host is empty or only contains a port
+			if u.Host == "" || strings.HasPrefix(u.Host, ":") {
+				return
+			}
+			if len(u.Query()) > 0 {
+				return
+			}
 		}
+		s.baseURL = strings.TrimSuffix(baseURL, "/")
 	}
 }
 
-func (s *SSEServer) setKeepAliveInterval(interval time.Duration) {
-	s.keepAlive = true
-	s.keepAliveInterval = interval
-}
-
-func (s *SSEServer) setKeepAlive(keepAlive bool) {
-	s.keepAlive = keepAlive
-}
-
-func (s *SSEServer) setContextFunc(fn HTTPContextFunc) {
-	s.contextFunc = fn
-}
-
-func (s *SSEServer) setHTTPServer(srv *http.Server) {
-	s.srv = srv
-}
-
-func (s *SSEServer) setBaseURL(baseURL string) {
-	s.baseURL = baseURL
+// WithStaticBasePath adds a new option for setting a static base path
+func WithStaticBasePath(basePath string) SSEOption {
+	return func(s *SSEServer) {
+		s.basePath = normalizeURLPath(basePath)
+	}
 }
 
 // WithBasePath adds a new option for setting a static base path.
@@ -148,11 +151,26 @@ func WithBasePath(basePath string) SSEOption {
 	return WithStaticBasePath(basePath)
 }
 
+// WithDynamicBasePath accepts a function for generating the base path. This is
+// useful for cases where the base path is not known at the time of SSE server
+// creation, such as when using a reverse proxy or when the server is mounted
+// at a dynamic path.
+func WithDynamicBasePath(fn DynamicBasePathFunc) SSEOption {
+	return func(s *SSEServer) {
+		if fn != nil {
+			s.dynamicBasePathFunc = func(r *http.Request, sid string) string {
+				bp := fn(r, sid)
+				return normalizeURLPath(bp)
+			}
+		}
+	}
+}
+
 // WithMessageEndpoint sets the message endpoint path
 func WithMessageEndpoint(endpoint string) SSEOption {
-	return sseOption(func(s *SSEServer) {
+	return func(s *SSEServer) {
 		s.messageEndpoint = endpoint
-	})
+	}
 }
 
 // WithAppendQueryToMessageEndpoint configures the SSE server to append the original request's
@@ -161,37 +179,53 @@ func WithMessageEndpoint(endpoint string) SSEOption {
 // SSE connection request and carry them over to subsequent message requests, maintaining
 // context or authentication details across the communication channel.
 func WithAppendQueryToMessageEndpoint() SSEOption {
-	return sseOption(func(s *SSEServer) {
+	return func(s *SSEServer) {
 		s.appendQueryToMessageEndpoint = true
-	})
+	}
 }
 
 // WithUseFullURLForMessageEndpoint controls whether the SSE server returns a complete URL (including baseURL)
 // or just the path portion for the message endpoint. Set to false when clients will concatenate
 // the baseURL themselves to avoid malformed URLs like "http://localhost/mcphttp://localhost/mcp/message".
 func WithUseFullURLForMessageEndpoint(useFullURLForMessageEndpoint bool) SSEOption {
-	return sseOption(func(s *SSEServer) {
+	return func(s *SSEServer) {
 		s.useFullURLForMessageEndpoint = useFullURLForMessageEndpoint
-	})
+	}
 }
 
 // WithSSEEndpoint sets the SSE endpoint path
 func WithSSEEndpoint(endpoint string) SSEOption {
-	return sseOption(func(s *SSEServer) {
+	return func(s *SSEServer) {
 		s.sseEndpoint = endpoint
-	})
+	}
+}
+
+// WithHTTPServer sets the HTTP server instance
+func WithHTTPServer(srv *http.Server) SSEOption {
+	return func(s *SSEServer) {
+		s.srv = srv
+	}
+}
+
+func WithKeepAliveInterval(keepAliveInterval time.Duration) SSEOption {
+	return func(s *SSEServer) {
+		s.keepAlive = true
+		s.keepAliveInterval = keepAliveInterval
+	}
+}
+
+func WithKeepAlive(keepAlive bool) SSEOption {
+	return func(s *SSEServer) {
+		s.keepAlive = keepAlive
+	}
 }
 
 // WithSSEContextFunc sets a function that will be called to customise the context
 // to the server using the incoming request.
-//
-// Deprecated: Use WithContextFunc instead. This will be removed in a future version.
-//
-//go:deprecated
 func WithSSEContextFunc(fn SSEContextFunc) SSEOption {
-	return sseOption(func(s *SSEServer) {
-		WithHTTPContextFunc(HTTPContextFunc(fn)).applyToSSE(s)
-	})
+	return func(s *SSEServer) {
+		s.contextFunc = fn
+	}
 }
 
 // NewSSEServer creates a new SSE server instance with the given MCP server and options.
@@ -207,15 +241,16 @@ func NewSSEServer(server *MCPServer, opts ...SSEOption) *SSEServer {
 
 	// Apply all options
 	for _, opt := range opts {
-		opt.applyToSSE(s)
+		opt(s)
 	}
 
 	return s
 }
 
-// NewTestServer creates a test server for testing purposes.
+// NewTestServer creates a test server for testing purposes
 func NewTestServer(server *MCPServer, opts ...SSEOption) *httptest.Server {
 	sseServer := NewSSEServer(server, opts...)
+
 	testServer := httptest.NewServer(sseServer)
 	sseServer.baseURL = testServer.URL
 	return testServer
